Moved storage api to its own module
