+++ ethereum/statetransition/build.gradle
@@ -10,6 +10,7 @@ dependencies {
   implementation project(':services:serviceutils')
   implementation project(':ssz')
   implementation project(':storage')
+  implementation project(':storage:api')
   implementation project(':util')
 
   implementation 'com.google.code.gson:gson'
