+++ services/beaconchain/build.gradle
@@ -14,6 +14,7 @@ dependencies {
   implementation project(':services:powchain')
   implementation project(':ssz')
   implementation project(':storage')
+  implementation project(':storage:api')
   implementation project(':sync')
   implementation project(':util')
   implementation project(':validator:api')
