+++ settings.gradle
@@ -26,6 +26,7 @@ include 'pow'
 include 'protoarray'
 include 'ssz'
 include 'storage'
+include 'storage:api'
 include 'sync'
 include 'validator:anticorruption'
 include 'validator:api'
