+++ storage/api/src/main/java/tech/pegasys/teku/storage/api/schema/ReplayDepositsResult.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.storage.server;
+package tech.pegasys.teku.storage.api.schema;
 
 import com.google.common.primitives.UnsignedLong;
 import java.util.Optional;
