+++ storage/src/main/java/tech/pegasys/teku/storage/server/DepositStorage.java
@@ -22,6 +22,7 @@ import tech.pegasys.teku.pow.api.Eth1EventsChannel;
 import tech.pegasys.teku.pow.event.DepositsFromBlockEvent;
 import tech.pegasys.teku.pow.event.MinGenesisTimeBlockEvent;
 import tech.pegasys.teku.storage.api.Eth1DepositStorageChannel;
+import tech.pegasys.teku.storage.api.schema.ReplayDepositsResult;
 import tech.pegasys.teku.util.async.SafeFuture;
 
 public class DepositStorage implements Eth1DepositStorageChannel, Eth1EventsChannel {
